+++ lib/gds_api/test_helpers/asset_manager.rb
@@ -129,6 +129,11 @@ module GdsApi
           .to_return(body: body.to_json, status: 200)
       end
 
+      def stub_asset_manager_delete_asset_forbidden(asset_id)
+        stub_request(:delete, "#{ASSET_MANAGER_ENDPOINT}/assets/#{asset_id}")
+          .to_return(status: 403)
+      end
+
       def stub_asset_manager_delete_asset_missing(asset_id)
         stub_request(:delete, "#{ASSET_MANAGER_ENDPOINT}/assets/#{asset_id}")
           .to_return(status: 404)